+++ src/plugins/printerdrivers/windows/qwindowsprintdevice.cpp
@@ -185,18 +185,18 @@ void QWindowsPrintDevice::loadPageSizes() const
    // Get the number of paper sizes and check all 3 attributes have same count
    DWORD paperCount = DeviceCapabilities(tmp.data(), NULL, DC_PAPERNAMES, NULL, NULL);
 
-   if (int(paperCount) > 0
-      && DeviceCapabilities(tmp.data(), NULL, DC_PAPERSIZE, NULL, NULL) == paperCount
-      && DeviceCapabilities(tmp.data(), NULL, DC_PAPERS, NULL, NULL) == paperCount) {
+   if ( static_cast<int>(paperCount) > 0
+         && DeviceCapabilities(tmp.data(), NULL, DC_PAPERSIZE, NULL, NULL) == static_cast<int>(paperCount)
+         && DeviceCapabilities(tmp.data(), NULL, DC_PAPERS, NULL, NULL) == static_cast<int>(paperCount)) {
 
       QScopedArrayPointer<wchar_t> paperNames(new wchar_t[paperCount * 64]);
       QScopedArrayPointer<POINT> winSizes(new POINT[paperCount * sizeof(POINT)]);
       QScopedArrayPointer<wchar_t> papers(new wchar_t[paperCount]);
 
       // Get the details and match the default paper size
-      if (DeviceCapabilities(tmp.data(), NULL, DC_PAPERNAMES, paperNames.data(), NULL) == paperCount
-         && DeviceCapabilities(tmp.data(), NULL, DC_PAPERSIZE, (wchar_t *)winSizes.data(), NULL) == paperCount
-         && DeviceCapabilities(tmp.data(), NULL, DC_PAPERS, papers.data(), NULL) == paperCount) {
+      if (DeviceCapabilities(tmp.data(), NULL, DC_PAPERNAMES, paperNames.data(), NULL) == static_cast<int>(paperCount)
+            && DeviceCapabilities(tmp.data(), NULL, DC_PAPERSIZE, (wchar_t *)winSizes.data(), NULL) == static_cast<int>(paperCount)
+            && DeviceCapabilities(tmp.data(), NULL, DC_PAPERS, papers.data(), NULL) == static_cast<int>(paperCount)) {
 
          // Returned size is in tenths of a millimeter
          const qreal multiplier = qt_pointMultiplier(QPageSize::Millimeter);
@@ -225,6 +225,7 @@ QPageSize QWindowsPrintDevice::defaultPageSize() const
 
    if (LPDEVMODE pDevMode = getDevmode(m_hPrinter, m_id)) {
       // Get the default paper size
+
       if (pDevMode->dmFields & DM_PAPERSIZE) {
          // Find the supported page size that matches, in theory default should be one of them
          for (const QPageSize &ps : m_pageSizes) {
@@ -234,6 +235,7 @@ QPageSize QWindowsPrintDevice::defaultPageSize() const
             }
          }
       }
+
       // Clean-up
       free(pDevMode);
    }
@@ -315,12 +317,13 @@ void QWindowsPrintDevice::loadResolutions() const
       QScopedArrayPointer<LONG> resolutions(new LONG[resCount * 2]);
       // Get the details and match the default paper size
 
-      if (DeviceCapabilities(tmp.data(), NULL, DC_ENUMRESOLUTIONS, (LPWSTR)resolutions.data(), NULL) == resCount) {
+      if (DeviceCapabilities(tmp.data(), NULL, DC_ENUMRESOLUTIONS, (LPWSTR)resolutions.data(), NULL) == static_cast<int>(resCount)) {
          for (int i = 0; i < int(resCount * 2); i += 2) {
             m_resolutions.append(resolutions[i + 1]);
          }
       }
    }
+
    m_haveResolutions = true;
 }
 
@@ -349,15 +352,15 @@ void QWindowsPrintDevice::loadInputSlots() const
 
    DWORD binCount = DeviceCapabilities(tmp.data(), NULL, DC_BINS, NULL, NULL);
 
-   if (int(binCount) > 0
-      && DeviceCapabilities(tmp.data(), NULL, DC_BINNAMES, NULL, NULL) == binCount) {
+   if (static_cast<int>(binCount) > 0
+      && DeviceCapabilities(tmp.data(), NULL, DC_BINNAMES, NULL, NULL) == static_cast<int>(binCount)) {
 
       QScopedArrayPointer<WORD> bins(new WORD[binCount * sizeof(WORD)]);
       QScopedArrayPointer<wchar_t> binNames(new wchar_t[binCount * 24]);
 
       // Get the details and match the default paper size
-      if (DeviceCapabilities(tmp.data(), NULL, DC_BINS, (LPWSTR)bins.data(), NULL) == binCount
-         && DeviceCapabilities(tmp.data(), NULL, DC_BINNAMES, binNames.data(), NULL) == binCount) {
+      if (DeviceCapabilities(tmp.data(), NULL, DC_BINS, (LPWSTR)bins.data(), NULL) == static_cast<int>(binCount)
+         && DeviceCapabilities(tmp.data(), NULL, DC_BINNAMES, binNames.data(), NULL) == static_cast<int>(binCount)) {
 
          for (int i = 0; i < int(binCount); ++i) {
             wchar_t *binName = binNames.data() + (i * 24);
@@ -379,6 +382,7 @@ QPrint::InputSlot QWindowsPrintDevice::defaultInputSlot() const
       // Get the default input slot
       if (pDevMode->dmFields & DM_DEFAULTSOURCE) {
          QPrint::InputSlot tempSlot = paperBinToInputSlot(pDevMode->dmDefaultSource, QString());
+
          for (const QPrint::InputSlot &slot : supportedInputSlots()) {
             if (slot.key == tempSlot.key) {
                inputSlot = slot;
@@ -411,6 +415,7 @@ void QWindowsPrintDevice::loadDuplexModes() const
       m_duplexModes.append(QPrint::DuplexLongSide);
       m_duplexModes.append(QPrint::DuplexShortSide);
    }
+
    m_haveDuplexModes = true;
 }
